Create Calculator.py
I have made like 3 or so calculators in python and have created saparate repos for them all. This is the new calculator which has some more features in it if you will read my old repo and the new one than you will get the idea that what are the new features which I had made in this.